FEAT(Validation.java) : Change data type

+++ src/main/java/lotto/Validation.java
@@ -5,19 +5,19 @@ import java.util.List;
 import static Comment.ErrorMessage.*;
 
 public class Validation {
-    public static void minPriceValidation(int price) {
+    public static void minPriceValidation(long price) {
         if (price < 1_000) {
             throw new IllegalArgumentException(MIN_COST_ERROR);
         }
     }
 
-    public static void maxPriceValidation(int price) {
+    public static void maxPriceValidation(long price) {
         if (price > 100_000) {
             throw new IllegalArgumentException(MAX_COST_ERROR);
         }
     }
 
-    public static void monetaryUnitValidation(int price) {
+    public static void monetaryUnitValidation(long price) {
         if (price % 1_000 != 0) {
             throw new IllegalArgumentException(MONETARY_UNIT_ERROR);
         }
